// src/app/model/use.model.ts
export class user {
  // id: string;
  firstName: string;
  lastName: string;
  email: string;
  mobile: Number;
  age: Number;
  country: string;
  state: string;
  homeAddress1: string;
  homeAddress2: string;
  companyAddress1: string;
  companyAddress2: string;
  // code: string;
  // contactNumber: number;
  // address: string;
}